requirements.txt
streamlit==1.32.2
torch==2.1.0
transformers==4.36.2
sentencepiece
protobuf<5
Pillow
google-generativeai
